refactor: move types to separate file
Create `types.ts` with type definitioans.
Add types for messages and chats.
